restructure MVC to handle bios and stats
