+++ controllers/playerBioController.js
@@ -1,0 +1,45 @@
+const db = require('../models')
+
+// Defining methods for the playerController
+module.exports = {
+
+  create: function (req, res) {
+    db.PlayerBios
+      .create(req.body)
+      .then(dbModel => res.json(dbModel))
+      .catch(err => res.status(422).json(err))
+  }
+
+//   findAll: function(req, res) {
+//     db.Book
+//       .find(req.query)
+//       .sort({ date: -1 })
+//       .then(dbModel => res.json(dbModel))
+//       .catch(err => res.status(422).json(err));
+//   },
+//   findById: function(req, res) {
+//     db.Book
+//       .findById(req.params.id)
+//       .then(dbModel => res.json(dbModel))
+//       .catch(err => res.status(422).json(err));
+//   },
+//   create: function(req, res) {
+//     db.Book
+//       .create(req.body)
+//       .then(dbModel => res.json(dbModel))
+//       .catch(err => res.status(422).json(err));
+//   },
+//   update: function(req, res) {
+//     db.Book
+//       .findOneAndUpdate({ _id: req.params.id }, req.body)
+//       .then(dbModel => res.json(dbModel))
+//       .catch(err => res.status(422).json(err));
+//   },
+//   remove: function(req, res) {
+//     db.Book
+//       .findById({ _id: req.params.id })
+//       .then(dbModel => dbModel.remove())
+//       .then(dbModel => res.json(dbModel))
+//       .catch(err => res.status(422).json(err));
+//   }
+}
